feat: Add refresh token authentication to AuthController
The code changes in `auth.controller.ts` add refresh token authentication to the `refresh` endpoint. The `@Auth` decorator is used to require authentication with the `user` role. The `@GetUser('_id')` decorator is used to retrieve the user ID from the access token. The `refreshToken` method in `auth.service.ts` is updated to accept the user ID as a parameter.

Recent repository commits:
- feat: Add flat option to generateOptions in nest-cli.json
- feat: Add private routes with authentication
- feat: Update auth.service.ts to use mongoose.Types.ObjectId for user ID
- chore: Remove console.log statement in AppModule constructor
- boilerplate nestjs mongodb